feat and refactor: nweet 기본 이미지 추가 & 파일모듈 분리
